Create MockCourseTestCase
