+++ tests/test_utils.py
@@ -4,8 +4,6 @@ from django.test import TestCase
 from django.http import Http404
 from django.core.exceptions import ValidationError
 
-from opaque_keys.edx.locator import CourseLocator, BlockUsageLocator
-
 # We need this in order to mock.patch get_course_by_id
 from edx_shopify import utils
 
@@ -16,7 +14,7 @@ from edx_shopify.utils import process_order, process_line_item
 
 from edx_shopify.models import Order
 
-from . import JsonPayloadTestCase
+from . import JsonPayloadTestCase, MockCourseTestCase
 
 try:
     from unittest.mock import Mock, patch
@@ -80,7 +78,7 @@ class ProcessOrderTest(JsonPayloadTestCase):
         self.assertEqual(order.status, Order.PROCESSING)
 
 
-class ProcessLineItemTest(TestCase):
+class ProcessLineItemTest(MockCourseTestCase):
 
     def test_invalid_line_item(self):
         order = Order()
@@ -116,37 +114,7 @@ class ProcessLineItemTest(TestCase):
                 process_line_item(order, line_item)
 
 
-class EmailEnrollmentTest(TestCase):
-
-    def setUp(self):
-        # Set up a mock course
-        course_id_string = 'course-v1:org+course+run1'
-        cl = CourseLocator.from_string(course_id_string)
-        bul = BlockUsageLocator(cl, u'course', u'course')
-        course = Mock()
-        course.id = cl
-        course.system = Mock()
-        course.scope_ids = Mock()
-        course.scope_id.user_id = None
-        course.scope_ids.block_type = u'course'
-        course.scope_ids.def_id = bul
-        course.scope_ids.usage_id = bul
-        course.location = bul
-        course.display_name = u'Course - Run 1'
-
-        self.course_id_string = course_id_string
-        self.cl = cl
-        self.course = course
-
-        email_params = {'registration_url': u'https://localhost:8000/register',  # noqa: E501
-                        'course_about_url': u'https://localhost:8000/courses/course-v1:org+course+run1/about',  # noqa: E501
-                        'site_name': 'localhost:8000',
-                        'course': course,
-                        'is_shib_course': None,
-                        'display_name': u'Course - Run 1',
-                        'auto_enroll': True,
-                        'course_url': u'https://localhost:8000/courses/course-v1:org+course+run1/'}  # noqa: E501
-        self.email_params = email_params
+class EmailEnrollmentTest(MockCourseTestCase):
 
     def test_enrollment_failure(self):
         # Enrolling in a non-existent course (or run) should fail, no
